requirements.txt
tensorflow
numpy
sklearn
coloredlogs
albumentations
tqdm
opencv-python
pillow
pandas
mlflow
segmentation_models